refactor: restaurants -> restaurantManager로 파일명 수정 및 import 경로 수정
