+++ src/domain/restaurantManager.ts
